Changing eslint config to 2 separate files, moving app.js to template code
