+++ commands/init/templates/src/app.js
