--- commands/init/templates/src/entry.js
@@ -1,3 +1,0 @@
-import app from './';
-
-app.render();
