+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mozilla-neo",
-  "version": "1.4.0",
+  "version": "2.0.0",
   "engines": {
     "node": ">=6.0.0"
   },
@@ -12,13 +12,10 @@
     "commands",
     "config"
   ],
-  "main": "build/app.js",
   "bin": {
     "neo": "./bin/neo"
   },
   "scripts": {
-    "build": "mkdir -p build && babel src/app.js --source-maps inline -o build/app.js && cp src/template.ejs build",
-    "prepublish": "npm run build",
     "test": "npm explore $(node fixture) -- 'npm install && npm test'"
   },
   "babel": {
@@ -94,7 +91,6 @@
     "exports-loader": "0.6.3",
     "file-loader": "0.9.0",
     "html-webpack-plugin": "2.22.0",
-    "immutable": "3.8.1",
     "imports-loader": "0.6.5",
     "isparta-instrumenter-loader": "1.0.1",
     "json-loader": "0.5.4",
@@ -110,20 +106,7 @@
     "mocha": "3.0.0",
     "node-sass": "3.8.0",
     "qs": "6.2.1",
-    "react": "15.3.0",
-    "react-dom": "15.3.0",
     "react-hot-loader": "1.3.0",
-    "react-redux": "4.4.5",
-    "react-router": "2.6.1",
-    "react-router-redux": "4.0.5",
-    "redux": "3.5.2",
-    "redux-devtools": "3.3.1",
-    "redux-devtools-chart-monitor": "1.4.2",
-    "redux-devtools-dock-monitor": "1.1.1",
-    "redux-devtools-log-monitor": "1.0.11",
-    "redux-immutable": "3.0.6",
-    "redux-logger": "2.6.1",
-    "redux-slider-monitor": "1.0.7",
     "sass-loader": "4.0.0",
     "style-loader": "0.13.1",
     "url-loader": "0.5.7",
@@ -136,7 +119,6 @@
     "yeoman-generator": "0.24.1"
   },
   "devDependencies": {
-    "babel-cli": "6.11.4",
     "yeoman-test": "1.4.0"
   },
   "bugs": {
